Added db persistence for commits.
Added a repository class for commit that also implements the commit service so it can stand in for the old github service or stub.
Added a new route /refresh that calls github service to get latest commits and persists them with commits repository.
Swapped controller injection of github commit service with commit repository for mvctopcommits. Now we're using db values instead!